rest_movie/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
contourpy==1.3.0
cryptography==43.0.0
cycler==0.12.1
defusedxml==0.8.0rc2
Django==5.1
django-ckeditor==6.7.1
django-filter==24.3
django-js-asset==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
fonttools==4.53.1
idna==3.8
kiwisolver==1.4.5
Markdown==3.7
matplotlib==3.9.2
numpy==2.1.0
oauthlib==3.2.2
packaging==24.1
pillow==10.4.0
pycparser==2.22
PyJWT==2.9.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==74.0.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
urllib3==2.2.2
